refactor: extract transaction grouping util
